Move errors to own module
