+++ p2panda-rs/src/test_utils/memory_store/mod.rs
@@ -6,6 +6,7 @@
 //! - `domain` and `validation` modules for publishing and validating entries and operations
 //! - helpers for populating a store with test data
 pub mod domain;
+pub mod errors;
 pub mod helpers;
 mod provider;
 mod stores;
